Refactor and move PreviewResultInputPlugin to org.embulk.test
