+++ embulk-test/src/main/java/org/embulk/test/TestingBulkLoader.java
@@ -10,25 +10,18 @@ import com.google.inject.Module;
 import com.google.inject.util.Modules;
 import java.util.List;
 
-import org.embulk.config.ConfigDiff;
 import org.embulk.config.ConfigSource;
-import org.embulk.config.Task;
 import org.embulk.config.TaskReport;
-import org.embulk.config.TaskSource;
 import org.embulk.exec.BulkLoader;
 import org.embulk.exec.ExecutionResult;
 import org.embulk.exec.ForSystemConfig;
-import org.embulk.exec.PreviewResult;
 import org.embulk.exec.ResumeState;
 import org.embulk.spi.Exec;
 import org.embulk.spi.ExecSession;
 import org.embulk.spi.InputPlugin;
-import org.embulk.spi.Page;
-import org.embulk.spi.PageOutput;
 import org.embulk.spi.Schema;
 import org.slf4j.Logger;
 
-import static com.google.common.base.Preconditions.checkState;
 import static org.embulk.plugin.InjectedPluginSource.registerPluginTo;
 
 class TestingBulkLoader
@@ -50,54 +43,6 @@ class TestingBulkLoader
                 return ImmutableList.of(Modules.override(modules).with(ImmutableList.of(override)));
             }
         };
-    }
-
-    /**
-     * This plugin is used for TestingEmbulk.InputBuilder.preview().
-     */
-    static final class PreviewResultInputPlugin
-            implements InputPlugin
-    {
-        private static PreviewResult previewResult;
-
-        static void setPreviewResult(PreviewResult result)
-        {
-            previewResult = result;
-        }
-
-        @Override
-        public ConfigDiff transaction(ConfigSource config, Control control)
-        {
-            checkState(previewResult != null, "PreviewResult object must be set");
-            return resume(config.loadConfig(Task.class).dump(), previewResult.getSchema(), 1, control);
-        }
-
-        @Override
-        public ConfigDiff resume(TaskSource taskSource, Schema schema, int taskCount, Control control)
-        {
-            control.run(taskSource, schema, taskCount);
-            return Exec.newConfigDiff();
-        }
-
-        @Override
-        public void cleanup(TaskSource taskSource, Schema schema, int taskCount, List<TaskReport> successTaskReports)
-        {
-        }
-
-        @Override
-        public TaskReport run(TaskSource taskSource, Schema schema, int taskIndex, PageOutput output)
-        {
-            for (Page page : previewResult.getPages()) {
-                output.add(page);
-            }
-            return Exec.newTaskReport();
-        }
-
-        @Override
-        public ConfigDiff guess(ConfigSource config)
-        {
-            return Exec.newConfigDiff();
-        }
     }
 
     @Inject
